requirements.txt
asgiref==3.4.1
beautifulsoup4==4.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
cycler==0.11.0
Cython==0.29.23
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gensim==4.1.2
idna==3.3
joblib==1.1.0
JPype1==1.3.0
kiwisolver==1.3.2
konlpy==0.5.2
lxml==4.6.4
matplotlib==3.4.3
numpy==1.21.4
oauthlib==3.1.1
pandas==1.3.4
Pillow==8.4.0
PyJWT==1.7.1
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sqlparse==0.4.2
threadpoolctl==3.0.0
tweepy==3.10.0
urllib3==1.26.7
